weekly

+++ src/main/java/com/example/review_test/question/ReviewPostRepository.java
@@ -2,5 +2,5 @@ package com.example.review_test.question;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ReviewPostRepository extends JpaRepository<ReviewPost, Integer> {
+public interface ReviewPostRepository extends JpaRepository<ReviewPost, Long> {
 }
